// src/api/generated/models/bannerItem.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Nest Shop API
 * API for my startup
 * OpenAPI spec version: 1.0.0
 */
import type { BannerItemStore } from './bannerItemStore';
import type { BannerProductEntry } from './bannerProductEntry';
import type { BannerSlot } from './bannerSlot';
import type { BannerStatus } from './bannerStatus';
import type { PromoType } from './promoType';

export interface BannerItem {
  id?: number;
  storeId?: number;
  title?: string;
  accent?: string;
  description?: string;
  decoNum?: string;
  promoTag?: string;
  color?: string;
  promoType?: PromoType;
  /** @nullable */
  discount?: number | null;
  /** @nullable */
  fixedPrice?: number | null;
  deadline?: string;
  isActive?: boolean;
  status?: BannerStatus;
  /** @nullable */
  rejectReason?: string | null;
  createdAt?: string;
  store?: BannerItemStore;
  /** @maxItems 3 */
  products?: BannerProductEntry[];
  slot?: BannerSlot;
}